feat: 토큰 갱신 로직 activity에 적용
